LPS-98654 Move configuration where it is used
There is no need to expose new API, as the configuration is only
referenced from one module.  Move everything there.
