+++ modules/apps/adaptive-media/adaptive-media-document-library-thumbnails/src/main/resources/content/Language_fi.properties
