+++ modules/apps/adaptive-media/adaptive-media-document-library-thumbnails/src/main/resources/content/Language_ta_IN.properties
